# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ActionNodeRead": {
        "properties": {
          "automation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Automation Id"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invocation_count": {
            "default": 0,
            "title": "Invocation Count",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "trigger",
              "action",
              "outcome"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutcomeRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "playbook_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playbook Id"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "sync_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Error"
          },
          "sync_status": {
            "enum": [
              "unprovisioned",
              "pending",
              "enabled",
              "disabled",
              "error"
            ],
            "title": "Sync Status",
            "type": "string"
          },
          "trigger": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriggerRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "kind",
          "position",
          "name",
          "playbook_id",
          "prompt",
          "enabled",
          "sync_status",
          "sync_error",
          "automation_id"
        ],
        "title": "ActionNodeRead",
        "type": "object"
      },
...
      "EdgeCreate": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/NodeRef"
          },
          "target": {
            "$ref": "#/components/schemas/NodeRef"
          }
        },
        "required": [
          "source",
          "target"
        ],
        "title": "EdgeCreate",
        "type": "object"
      },
      "EdgeRead": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "outcome_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Count"
          },
          "source": {
            "$ref": "#/components/schemas/NodeRef"
          },
          "target": {
            "$ref": "#/components/schemas/NodeRef"
          }
        },
        "required": [
          "id",
          "source",
          "target"
        ],
        "title": "EdgeRead",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutcomeUpdate"
              },
              {
                "type": "null"
              }
            ]
          },
          "playbook_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playbook Id"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          },
          "prompt": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "trigger": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriggerUpdate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "position"
        ],
        "title": "NodeCreate",
        "type": "object"
      },
      "NodeRead": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "trigger",
              "action",
              "outcome"
            ],
            "title": "Kind",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutcomeRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          },
          "trigger": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriggerRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "kind",
          "position"
        ],
        "title": "NodeRead",
        "type": "object"
      },
      "NodeRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "trigger",
              "action",
              "outcome"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind"
        ],
        "title": "NodeRef",
        "type": "object"
      },
      "NodeUpdate": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutcomeUpdate"
              },
              {
                "type": "null"
              }
            ]
          },
          "playbook_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playbook Id"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Position"
              },
              {
                "type": "null"
              }
            ]
          },
          "prompt": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "trigger": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriggerUpdate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "NodeUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "devin-flow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/canvas/edges": {
      "post": {
        "operationId": "create_edge_api_canvas_edges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EdgeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas object not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Edge"
      }
    },
    "/api/canvas/edges/{edge_id}": {
      "delete": {
        "operationId": "delete_edge_api_canvas_edges__edge_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "edge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas object not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Edge"
      }
    },
    "/api/canvas/nodes/{kind}": {
      "post": {
        "operationId": "create_node_api_canvas_nodes__kind__post",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "trigger",
                "action",
                "outcome"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ActionNodeRead"
                    },
                    {
                      "$ref": "#/components/schemas/NodeRead"
                    }
                  ],
                  "title": "Response Create Node Api Canvas Nodes  Kind  Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas object not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Node"
      }
    },
    "/api/canvas/nodes/{kind}/{node_id}": {
      "delete": {
        "operationId": "delete_node_api_canvas_nodes__kind___node_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "trigger",
                "action",
                "outcome"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas object not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Node"
      },
      "patch": {
        "operationId": "update_node_api_canvas_nodes__kind___node_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "trigger",
                "action",
                "outcome"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ActionNodeRead"
                    },
                    {
                      "$ref": "#/components/schemas/NodeRead"
                    }
                  ],
                  "title": "Response Update Node Api Canvas Nodes  Kind   Node Id  Patch"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas object not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Canvas conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Node"
      }
    },
...
  }
}